requirements.txt
watchdog
aiohttp
numpy
transformers
torchvision
onnx
onnxruntime-training-cpu
torch
evaluate
selenium
scikit-learn
